+++ app/src/main/java/io/gnosis/safe/di/modules/ViewModule.kt
@@ -23,7 +23,7 @@ import io.gnosis.safe.ui.settings.SettingsViewModel
 import io.gnosis.safe.ui.settings.app.AppSettingsViewModel
 import io.gnosis.safe.ui.settings.app.GetInTouchViewModel
 import io.gnosis.safe.ui.settings.app.fiat.AppFiatViewModel
-import io.gnosis.safe.ui.settings.owner.OwnerNameViewModel
+import io.gnosis.safe.ui.settings.owner.OwnerEnterNameViewModel
 import io.gnosis.safe.ui.settings.owner.OwnerSeedPhraseViewModel
 import io.gnosis.safe.ui.settings.owner.list.OwnerListViewModel
 import io.gnosis.safe.ui.settings.owner.selection.OwnerSelectionViewModel
@@ -118,7 +118,7 @@ class ViewModule(
 
     @Provides
     @ForView
-    fun providesOwnerNameViewModel(provider: ViewModelProvider) = provider[OwnerNameViewModel::class.java]
+    fun providesOwnerNameViewModel(provider: ViewModelProvider) = provider[OwnerEnterNameViewModel::class.java]
 
     @Provides
     @ForView
