+++ app/src/main/java/io/gnosis/safe/ui/settings/owner/OwnerEditNameFragment.kt
@@ -1,0 +1,4 @@
+package io.gnosis.safe.ui.settings.owner
+
+class OwnerEditNameFragment {
+}
